Move get_location to utils
